Move slider components to components/slider
