+++ src/App.svelte
@@ -19,10 +19,10 @@
   import { onMount } from 'svelte';
 
   // Components
-  import Billboard from './components/Billboard.svelte';
-  import GlobalModal from './components/GlobalModal.svelte';
-  import Nav from './components/nav/Nav.svelte';
-  import Row from './components/Row.svelte';
+  import Billboard from '@components/Billboard.svelte';
+  import GlobalModal from '@components/GlobalModal.svelte';
+  import Nav from '@components/nav/Nav.svelte';
+  import Row from '@components/slider/Row.svelte';
 
   // Stores
   import { continueWatching, error, initializeMedia } from '@stores';
