+++ src/components/slider/Slider.svelte
