+++ src/components/slider/SliderPaginationIndicators.svelte
